--- app/diagnosis/_layout.tsx
@@ -1,5 +1,0 @@
-import { Slot } from "expo-router";
-
-export default function Layout() {
-  return <Slot />;
-}
